Integrate with Volto v9's new AppExtras component
