+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Not yet released
+
+- integration with the new `AppExtras` component in Volto v9.2 @silviubogan
+
 ## 0.1.0
 
 - Initial release
