+++ src/index.js
@@ -1,7 +1,16 @@
 import { matomo } from './reducers';
+import { MatomoAppExtra } from './MatomoAppExtra';
 
 const applyConfig = (config) => {
   config.addonReducers.matomo = matomo;
+
+  config.settings.appExtras = [
+    ...config.settings.appExtras,
+    {
+      match: '',
+      component: MatomoAppExtra,
+    },
+  ];
 
   return config;
 };
